requirements.txt
Flask==3.0.3
tensorflow-gpu==2.10.0
pandas==2.0.3
numpy==1.24.4
h5py==3.11.0
Werkzeug==3.0.6
Jinja2==3.1.4
MarkupSafe==2.1.5
scikit-learn==1.3.2 